html/browsers/browsing-the-web/back-forward-cache/pushstate.https.html

<!DOCTYPE HTML>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script>
// See https://github.com/whatwg/html/issues/6207 for discussion on the
// specified, implemented and desired behaviors.
for (const bfcacheDisabled of [false, true]) {
  runBfcacheTest({
    funcBeforeNavigation: async (bfcacheDisabled) => {
      const urlPushState = location.href + '&pipe=sub';
      if (bfcacheDisabled) {
        await disableBFCache();
      }

      // `pushState(..., urlPushState)` on `urlA`,
      history.pushState('blue', '', urlPushState);
    },
    argsBeforeNavigation: [bfcacheDisabled],
    shouldBeCached: !bfcacheDisabled,
    funcAfterAssertion: async (pageA) => {
      const urlA = location.origin + executorPath + pageA.context_id;
      const urlPushState = urlA + '&pipe=sub';

      // and then navigate to `urlB` and then back navigate
      // (already done within `runBfcacheTest()`).
      // After the back navigation, `location` etc. should point to
      // `urlPushState` and the state that's pushed.
      assert_equals(await pageA.execute_script(() => location.href),
                    urlPushState, 'url');
      assert_equals(await pageA.execute_script(() => history.state),
                    'blue', 'history.state');

      if (bfcacheDisabled) {
        // When the page is not restored from BFCache, the HTML page is loaded
        // from `urlPushState` (not from `urlA`).
        // To detect whether the page is loaded from `urlA` or `urlPushState`,
        // we check whether `pipe=sub` is enabled on the executor.
        const loadedFromUrlPushState =
            await pageA.execute_script(() => hasPipeSub);
        assert_true(loadedFromUrlPushState,
            'document should be loaded from urlPushState with pipe=sub');
      }
    }
  }, 'back navigation to pushState()d page (' +
     (bfcacheDisabled ? 'not ' : '') + 'in BFCache)');
}
</script>
